Separate journal classes out a bit.
