+++ backend/src/jv2backend/tests/test_journalCollection.py
@@ -1,7 +1,8 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 # Copyright (c) 2023 Team JournalViewer and contributors
 
-from jv2backend.journals import JournalCollection, JournalFile, JournalData
+from jv2backend.journals import JournalCollection
+from jv2backend.journalFile import JournalFile, JournalData
 import xml.etree.ElementTree as ElementTree
 import datetime
 import pytest
