+++ backend/src/jv2backend/tests/test_journalData.py
@@ -1,10 +1,9 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 # Copyright (c) 2023 Team JournalViewer and contributors
 
-from jv2backend.journals import JournalData
+from jv2backend.journalFile import JournalData
 import jv2backend.select
 import xml.etree.ElementTree as ElementTree
-import json
 import pytest
 
 # Load test run data
